refactor

+++ Back-end/Controller/blog-controller.js
@@ -75,17 +75,19 @@ const getBlogById = async (req, res, next) => {
 /* Delete blog */
 const deleteBlog = async (req, res, next) => {
   /* how to delete blog from both the blog table and user table */
-  const id = req.params.id;
-  let blog;
-  try {
-    blog = await Blog.findByIdAndRemove(id).populate("user");
-    /* remove the blog from the user Array */
-    await blog.user.blogs.pull(blog);
-    await blog.user.save();
-  } catch (error) {
-    return console.log(error);
-  }
-  if (!blog) {
+   const id = req.params.id;
+   let blog
+   if(req.user.id===req.params.id||req.user.isAdmin){
+    try {
+      blog = await Blog.findByIdAndRemove(id).populate("user");
+      /* remove the blog from the user Array */
+      await blog.user.blogs.pull(blog);
+      await blog.user.save();
+    } catch (error) {
+      return console.log(error);
+    }
+   } 
+  else if (!blog) {
     return res.status(500).json({ message: "you can't delete blog" });
   }
   return res.status(200).json({ message: "blog deleted successfully" });
